+++ packages/bpk-docs/src/pages/AlignmentPage/AlignmentPage.js
@@ -24,15 +24,18 @@ import BpkLink from 'bpk-component-link';
 import BpkButton from 'bpk-component-button';
 import BpkText from 'bpk-component-text';
 import LongArrowRightIcon from 'bpk-component-icon/lg/long-arrow-right';
+import LongArrowRightIconSm from 'bpk-component-icon/sm/long-arrow-right';
+import AwardIcon from 'bpk-component-icon/lg/award';
 import DocsPageBuilder from './../../components/DocsPageBuilder';
 import Heading from './../../components/Heading';
 import Paragraph from './../../components/Paragraph';
 import * as ROUTES from './../../constants/routes';
 
-const AlignedBaseArrow = withAlignment(LongArrowRightIcon, lineHeightBase, iconSizeSm);
+const AlignedBaseArrow = withAlignment(LongArrowRightIconSm, lineHeightBase, iconSizeSm);
 const AlignedLargeArrow = withAlignment(LongArrowRightIcon, lineHeightLg, iconSizeLg);
-const AlignedButtonArrow = withButtonAlignment(LongArrowRightIcon);
+const AlignedButtonArrow = withButtonAlignment(LongArrowRightIconSm);
 const AlignedLargeButtonArrow = withLargeButtonAlignment(LongArrowRightIcon);
+const AlignedSpan = withAlignment('span', iconSizeLg, lineHeightBase);
 
 const components = [
   {
@@ -48,7 +51,7 @@ const components = [
         Using the HOC is done as follows:
       </Paragraph>,
       <BpkCodeBlock >
-        {`import LongArrowRightIcon from 'bpk-component-icon/lg/long-arrow-right';
+        {`import LongArrowRightIconSm from 'bpk-component-icon/sm/long-arrow-right';
 import { withAlignment } from 'bpk-component-icon';
 import { lineHeightBase, iconSizeSm } from 'bpk-tokens/tokens/base.es6';
 import BpkText from 'bpk-component-text';
@@ -107,22 +110,60 @@ const AlignedArrow = withAlignment(
     ],
   },
   {
+    id: 'withIcon',
+    title: 'Text with textStyle "base" alignment with an icon',
+    blurb: [
+      <Paragraph>
+        If wishing to use an icon taller than the text, alignment must be performed on the text as follows:
+      </Paragraph>,
+      <BpkCodeBlock >
+        {`import AwardIcon from 'bpk-component-icon/lg/award';
+import { withAlignment } from 'bpk-component-icon';
+import { iconSizeLg, lineHeightBase } from 'bpk-tokens/tokens/base.es6';
+import BpkText from 'bpk-component-text';
+
+const AlignedSpan = withAlignment(
+  'span', iconSizeLg, lineHeightBase
+);
+
+<BpkText textStyle="base" >
+  <AwardIcon fill={colors.colorGray700} />
+  <AlignedSpan >
+    &nbsp;
+    Search
+  </AlignedSpan>
+</BpkText>
+`}
+      </BpkCodeBlock>,
+    ],
+    examples: [
+      <BpkText textStyle="base" >
+        <AwardIcon fill={colors.colorGray700} />
+        <AlignedSpan >
+          &nbsp;
+          Search
+        </AlignedSpan>
+      </BpkText>,
+    ],
+  },
+  {
     id: 'withinButton',
     title: 'Icon alignment within a button',
     blurb: [
       <Paragraph>
-         HOC wrappers exist for common alignment tasks, such as aligning an icon within a button.
-         Other examples can be seen in the <BpkLink href={ROUTES.BUTTONS}>Buttons page</BpkLink>.
+        HOC wrappers exist for common alignment tasks, such as aligning an icon within a button.
+        Other examples can be seen in the <BpkLink href={ROUTES.BUTTONS}>Buttons page</BpkLink>.
       </Paragraph>,
       <BpkCodeBlock>
-        {`import LongArrowRightIcon from 'bpk-component-icon/lg/long-arrow-right';
+        {`import LongArrowRightIconSm from 'bpk-component-icon/sm/long-arrow-right';
 import { withButtonAlignment } from 'bpk-component-icon';
 
-const AlignedArrow = withButtonAlignment(LongArrowRightIcon);
+const AlignedArrow = withButtonAlignment(LongArrowRightIconSm);
 
 <BpkButton >
   Button Text
-  <AlignedArrow />
+  &nbsp;
+  <AlignedArrow fill={colors.colorWhite} />
 </BpkButton>
 `}
       </BpkCodeBlock>,
@@ -140,7 +181,7 @@ const AlignedArrow = withButtonAlignment(LongArrowRightIcon);
     title: 'Icon alignment within a large button',
     blurb: [
       <Paragraph>
-         Similarly, a HOC exists for aligning icons to a large button.
+        Similarly, a HOC exists for aligning icons to a large button.
          Other examples can be seen in the <BpkLink href={ROUTES.BUTTONS}>Buttons page</BpkLink>.
           </Paragraph>,
       <BpkCodeBlock>
@@ -149,15 +190,16 @@ import { withLargeButtonAlignment } from 'bpk-component-icon';
 
 const AlignedArrow = withLargeButtonAlignment(LongArrowRightIcon);
 
-<BpkButton >
+<BpkButton large >
   Button Text
+  &nbsp;
   <AlignedArrow />
 </BpkButton>
 `}
       </BpkCodeBlock>,
     ],
     examples: [
-      <BpkButton large>
+      <BpkButton large >
         Button Text
         &nbsp;
         <AlignedLargeButtonArrow fill={colors.colorWhite} />